pyproject.toml
[tool.poetry]
name = "mysite"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = ">=3.0,<3.1"
wagtail = ">=2.9,<2.10"
gunicorn = "^20.0.4"
django-filter = "^2.3.0"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
pep8 = "^1.7.1"
flake8 = "^3.8.3"
django-debug-toolbar = "^2.2"
ipython = "^7.16.1"
ipdb = "^0.13.3"

[tool.black]
line-length = 79
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
